修复android更新系统后创建数据文件夹失败的bug

+++ Mobile/Assets/Scripts/Runtime/LotteryManager.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using System.IO;
 using LotteryAnalyze;
+using System;
 
 public class LotteryManager : MonoBehaviour
 {
@@ -59,14 +60,31 @@ public class LotteryManager : MonoBehaviour
 
     void OnLog(string condition, string stackTrace, LogType type)
     {
-        if(LOG_FI == null)
+        if(LOG_FI == null || LOG_STREAM_WRITER == null)
         {
             LOG_FI = new FileInfo(LOG_PATH);
-            LOG_STREAM_WRITER = LOG_FI.CreateText();
+            try
+            {
+                if (File.Exists(LOG_PATH))
+                {
+                    LOG_STREAM_WRITER = LOG_FI.CreateText();
+                }
+                else
+                {
+                    LOG_STREAM_WRITER = new StreamWriter(LOG_PATH);
+                }
+            }
+            catch(Exception e)
+            {
+                Console.Write("Create Log file StreamWriter failed! - " + e.ToString());
+            }
         }
-        string msg = Time.time + " - " + type.ToString() + " : " + condition + "\nStack : \n" + stackTrace + "\n";
-        LOG_STREAM_WRITER.Write(msg);
-        LOG_STREAM_WRITER.Flush();
+        if (LOG_STREAM_WRITER != null)
+        {
+            string msg = Time.time + " - " + type.ToString() + " : " + condition + "\nStack : \n" + stackTrace + "\n";
+            LOG_STREAM_WRITER.Write(msg);
+            LOG_STREAM_WRITER.Flush();
+        }
     }
 
     string LOG_PATH = "";
@@ -78,25 +96,31 @@ public class LotteryManager : MonoBehaviour
     {
         LotteryAnalyze.AutoUpdateUtil.sCallBackOnCollecting += CallBack_CollectResult;
 
-#if UNITY_EDITOR_WIN || UNITY_EDITOR_WIN
-        LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER = Application.persistentDataPath + "/LotteryAnalyze";
-#elif UNITY_ANDROID
-        LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER = "/mnt/sdcard/LotteryAnalyze";
-#endif
+        //        string persistentDataPath = Application.persistentDataPath;
+        //#if UNITY_EDITOR_WIN || UNITY_EDITOR_WIN
+        //        LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER = Application.persistentDataPath + "/LotteryAnalyze";
+        //#elif UNITY_ANDROID
+        //        LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER = "/mnt/sdcard/LotteryAnalyze";
+        //#endif
+        //        LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER = Application.persistentDataPath + "/LotteryAnalyze";
+        LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER = GlobalSetting.ROOT_FOLDER;
 
         Debug.Log(LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER);
         if (!Directory.Exists(LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER))
             Directory.CreateDirectory(LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER);
 
-        Application.logMessageReceived += OnLog;
         LOG_PATH = LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER + "/log.txt";
         LOG_PATH = LOG_PATH.Replace('\\', '/');
-        File.Delete(LOG_PATH);
+        Application.logMessageReceived += OnLog;
+        if(File.Exists(LOG_PATH))
+            File.Delete(LOG_PATH);
+        Debug.Log("Create Log File : " + LOG_PATH);
 
         LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER += "/Data/";
         if (!Directory.Exists(LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER))
             Directory.CreateDirectory(LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER);
-        
+        Debug.Log("Create Data Folder : " + LotteryAnalyze.AutoUpdateUtil.DATA_PATH_FOLDER);
+
         LotteryAnalyze.GlobalSetting.ReadCfg();
     }
